fix(balances): refresh balances on lending pools and swaps
