+++ src/pages/Lending/LendingIndex.tsx
@@ -1,28 +1,37 @@
 import React, { useMemo } from 'react';
-import { ScrollView } from 'react-native';
-import { SafeAreaPage } from 'templates/SafeAreaPage';
+import { RefreshControl } from 'react-native';
+import { PageContainer, SafeAreaPage } from 'templates/SafeAreaPage';
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
-import { globalStyles } from 'global.styles';
 import { LendingRoutesStackProps } from 'routers/lending.routes';
 import { currentChainId } from 'utils/helpers';
 import { lendingTokens } from 'config/lending-tokens';
 import { LendingPool } from './components/LendingPool';
+import { useWalletAddress } from 'hooks/useWalletAddress';
+import { useGlobalLoans } from 'hooks/app-context/useGlobalLoans';
 
 type Props = NativeStackScreenProps<LendingRoutesStackProps, 'lending.index'>;
 
 export const LendingIndex: React.FC<Props> = () => {
   const chainId = currentChainId();
+  const owner = useWalletAddress()?.toLowerCase();
+  const { execute } = useGlobalLoans(owner);
   const pools = useMemo(
     () => lendingTokens.filter(item => item.chainId === chainId),
     [chainId],
   );
   return (
-    <SafeAreaPage>
-      <ScrollView style={globalStyles.page}>
+    <SafeAreaPage
+      scrollView
+      scrollViewProps={{
+        refreshControl: (
+          <RefreshControl refreshing={false} onRefresh={execute} />
+        ),
+      }}>
+      <PageContainer>
         {pools.map(item => (
           <LendingPool key={item.loanTokenAddress} lendingToken={item} />
         ))}
-      </ScrollView>
+      </PageContainer>
     </SafeAreaPage>
   );
 };
